requirements.txt
google-cloud-aiplatform>=1.38.0
google-genai>=1.0.0
google-cloud-storage>=2.13.0
yt-dlp>=2024.1.0
python-docx>=1.1.0
python-dotenv>=1.0.0
rich>=13.0.0
reportlab>=4.0.0
json-repair>=0.25.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
